chore: create `unready_blocks.go`
